Adjust digit function and DOM variables
Calculator always displays a zero. I didn't want the user
to be able to click an operator before a number was selected. Digits function now checks for a number of conditions before it replaces the default value: zero. So now when the user clicks an operator first, it will apply the operation to the default value: zero.

Another function checkForOperators was added because I was using the logic in multiple functions.